requirements.txt
Django==4.2.11
psycopg2-binary
gunicorn
python-dotenv
django-silk==5.0.3
setuptools==68.2.2
pip==24.0
django-ninja
python-jose
passlib[bcrypt]